requirements.txt
pandas==0.23.4
dash==0.40.0
dash-core-components==0.47.0
dash-html-components==0.16.0
dash-daq==0.1.4
dash-table==3.6.0
six==1.12.0
PyGithub==1.43.7
ruamel.yaml==0.15.94